Remove logging from monitoring to custom processor
