+++ core/src/main/resources/SpringConfigurationContext.xml
@@ -88,58 +88,66 @@
 
 		<bean
 			name="pipeProcessor,inputValidatorProcessor,outputValidatorProcessor,inputWrapperProcessor,outputWrapperProcessor"
-			class="org.frankframework.processors.MonitoringPipeProcessor"
+			class="org.frankframework.processors.LogPipeProcessor"
 			autowire="byName"
 			scope="prototype"
 			>
 			<property name="pipeProcessor">
-				<!-- allows a managed exception to be stored in the session -->
 				<bean
-					class="org.frankframework.processors.InputOutputPipeProcessor"
+					class="org.frankframework.processors.MonitoringPipeProcessor"
 					autowire="byName"
 					scope="prototype"
 					>
 					<property name="pipeProcessor">
+						<!-- allows a managed exception to be stored in the session -->
 						<bean
-							class="org.frankframework.processors.ExceptionHandlingPipeProcessor"
+							class="org.frankframework.processors.InputOutputPipeProcessor"
 							autowire="byName"
 							scope="prototype"
 							>
 							<property name="pipeProcessor">
 								<bean
-									class="org.frankframework.processors.LimitingParallelExecutionPipeProcessor"
+									class="org.frankframework.processors.ExceptionHandlingPipeProcessor"
 									autowire="byName"
 									scope="prototype"
 									>
 									<property name="pipeProcessor">
 										<bean
-											class="org.frankframework.processors.TransactionAttributePipeProcessor"
+											class="org.frankframework.processors.LimitingParallelExecutionPipeProcessor"
 											autowire="byName"
 											scope="prototype"
 											>
 											<property name="pipeProcessor">
 												<bean
-													class="org.frankframework.processors.LockerPipeProcessor"
+													class="org.frankframework.processors.TransactionAttributePipeProcessor"
 													autowire="byName"
 													scope="prototype"
 													>
 													<property name="pipeProcessor">
 														<bean
-															class="org.frankframework.processors.CheckMessageSizePipeProcessor"
+															class="org.frankframework.processors.LockerPipeProcessor"
 															autowire="byName"
 															scope="prototype"
 															>
 															<property name="pipeProcessor">
 																<bean
-																	class="org.frankframework.processors.TrackPreviousPipeInMetadataProcessor"
+																	class="org.frankframework.processors.CheckMessageSizePipeProcessor"
 																	autowire="byName"
-																	scope="prototype">
+																	scope="prototype"
+																	>
 																	<property name="pipeProcessor">
 																		<bean
-																			class="org.frankframework.processors.CorePipeProcessor"
+																			class="org.frankframework.processors.TrackPreviousPipeInMetadataProcessor"
 																			autowire="byName"
-																			scope="prototype"
-																		/>
+																			scope="prototype">
+																			<property name="pipeProcessor">
+																				<bean
+																					class="org.frankframework.processors.CorePipeProcessor"
+																					autowire="byName"
+																					scope="prototype"
+																				/>
+																			</property>
+																		</bean>
 																	</property>
 																</bean>
 															</property>
